89 performance improve

+++ KioChess/Engine/Interfaces/IAttackEvaluationService.cs
@@ -6,7 +6,7 @@ namespace Engine.Interfaces
     public interface IAttackEvaluationService
     {
         void Initialize(BitBoard[] boards);
-        int StaticExchange(AttackBase attack);
+        short StaticExchange(AttackBase attack);
         void SetBoard(IBoard board);
     }
 }
